Add files via upload
Broad Institute of MIT and Harvard and NEU AI Skunkworks Mouse Brain Challenge 


The Klarman Cell Observatory (KCO), led by the Chair of the Faculty Aviv Regev, in collaboration with the Macosko’s lab (Stanley Center) at the Broad Institute of MIT and Harvard and NEU AI Skunkworks is launching a pedagogical machine learning challenge to recognize the view in slices of the adult mouse brain (M. musculus P56). The purpose of the challenge is to pre-screen candidates interested in working as Associate Computational Biologist (RA-equivalent) in the Regev lab. 


For challenge details see https://www.skonks.com/post/broad-institute-of-mit-and-harvard-and-neu-ai-skunkworks-mouse-brain-challenge